backend/pyproject.toml
[tool.poetry]
name = "reader"
version = "0.1.0"
description = ""
authors = ["Andrey Chernikov"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.75.2"
pydantic = "^1.9.0"
pymorphy2 = "^0.9.1"
pymorphy2-dicts-ru = "^2.4.417127"
uvicorn = "^0.17.6"
nltk = "^3.7"
tortoise-orm = "^0.19.2"
aerich = "^0.7.1"
asyncpg = "^0.27.0"

[tool.poetry.dev-dependencies]
flake8 = "^4.0.1"
isort = "^5.10.1"


[tool.aerich]
tortoise_orm = "config.db.TORTOISE_ORM"
location = "./migrations/"
src_folder = "./."
[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
